Use float for the degrees-to-radians conversion, too.
It seems like for at leasts some instruction set this is not fully expanded at build time.

+++ unpaper.h
@@ -239,7 +239,7 @@ void saveDebug(char *filenameTemplate, int index, AVFrame *image)
 
 /* --- arithmetic tool functions ------------------------------------------ */
 
-static inline double degreesToRadians(double d) {
+static inline float degreesToRadians(float d) {
     return d * M_PI / 180.0;
 }
